Search

+++ app/src/main/java/com/example/android/mynotenew/NoteViewModel.kt
@@ -43,8 +43,12 @@ class NoteViewModel(private val repository: NoteRepository) : ViewModel() {
     fun insert(note: NoteEntity) = viewModelScope.launch {
         repository.insert(note)
     }
-    fun delete(note: NoteEntity) = viewModelScope.launch {
-        repository.delete(note)
+    fun delete(position: Int) = viewModelScope.launch {
+        val note = allNotes.value?.get(position)
+        if (note!==null){
+            repository.delete(note)
+        }
+
     }
 }
